requirements.txt
absl-py==1.0.0
accelerate==0.20.3
adamod==0.0.3
addict==2.4.0
aiocontextvars==0.2.2
anykeystore==0.2
apex==0.1
appdirs==1.4.4
asn1crypto==0.22.0
astunparse==1.6.3
attrs==21.4.0
av==10.0.0
backcall==0.1.0
bleach==3.0.2
boto==2.49.0
botocore==1.11.6
cached-property==1.5.2
cachetools==4.1.1
catalyst==22.4
certifi==2017.7.27.1
cffi==1.14.0
chardet==3.0.4
charset-normalizer==2.0.12
chumpy==0.70
click==8.1.3
cloudpickle==0.6.1
colorama==0.4.6
commonmark==0.9.1
conda-pack==0.5.0
ConfigArgParse==1.5.3
contextvars==2.4
cryptacular==1.6.2
cryptography==37.0.4
cxxfilt==0.3.0
cycler==0.10.0
Cython==0.29.6
cython-bbox==0.1.3
dask==0.20.2
dataclasses==0.6
decorator==4.3.0
decord==0.6.0
defusedxml==0.7.1
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.14
dotmap==1.3.30
dpcpp-cpp-rt==2022.1.0
easydict==1.8
einops==0.6.0
entrypoints==0.2.3
et-xmlfile==1.1.0
filelock==3.12.2
flatbuffers==23.3.3
fonttools==4.33.3
fsspec==2023.1.0
ftfy==6.1.1
future==0.18.2
fvcore==0.1.3.post20210311
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
glog==0.3.1
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.33.2
h5py==3.1.0
halpecocotools==0.0.0
huggingface-hub==0.16.4
hupper==1.12
hydra-slayer==0.4.1
idna==2.6
imageio==2.17.0
immutables==0.14
importlib-metadata==3.1.0
iniconfig==1.1.1
install==1.3.5
intel-cmplr-lib-rt==2022.1.0
intel-cmplr-lic-rt==2022.1.0
intel-opencl-rt==2022.1.0
intel-openmp==2022.1.0
iopath==0.1.4
ipykernel==5.1.0
ipython==7.1.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.1
Jinja2==2.10
jmespath==0.9.4
joblib==0.17.0
jsonpatch==1.32
jsonpickle==1.5.2
jsonpointer==2.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.0.1
kornia==0.6.5
libclang==14.0.6
loguru==0.5.3
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.5.3
mediapipe==0.9.0.1
mistune==0.8.3
mkl==2022.1.0
mkl-fft==1.3.1
mkl-random==1.2.1
mkl-service==2.3.0
mmaction==0.5.0
mmcv==1.4.0
mmcv-full==1.3.16
mmdet==2.26.0
model-index==0.1.11
msgpack==1.0.0
msgpack-numpy==0.4.7.1
multimethod==1.4
munch==2.5.0
munkres==1.1.4
natsort==8.2.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.2
nms==0.1.6
nms-1d-cpu==0.0.0
notebook==5.7.2
numpy==1.21.6
oauthlib==3.1.0
opencv-contrib-python==3.4.10.37
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opendr==0.78
openpyxl==3.1.2
opt-einsum==3.3.0
ordered-set==4.1.0
packaging==22.0
pandas==1.3.5
pandocfilters==1.4.2
parso==0.3.1
PasteDeploy==3.0.1
pathtools==0.1.2
pbkdf2==1.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==9.5.0
pip==22.3.1
plaster==1.1.2
plaster-pastedeploy==1.0.1
pluggy==1.0.0
portalocker==2.2.1
prometheus-client==0.4.2
prompt-toolkit==2.0.7
protobuf==3.14.0
psutil==5.7.3
ptflops==0.6.9
ptyprocess==0.6.0
py==1.11.0
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pycparser==2.21
Pygments==2.13.0
pyOpenSSL==17.2.0
pyparsing==2.2.2
pyramid==2.0.1
pyramid-mailer==0.15.1
PySocks==1.6.7
pytest==7.1.2
python-dateutil==2.8.0
python-gflags==3.1.2
python3-openid==3.2.0
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==17.1.2
qtconsole==4.4.2
RandAugment==0.1
regex==2023.6.3
repoze.sendmail==4.4.1
requests==2.27.1
requests-oauthlib==1.3.0
rich==12.6.0
rsa==3.4.2
s3transfer==0.1.13
sacred==0.8.2
safetensors==0.4.1
scikit-image==0.19.2
scikit-learn==0.23.2
scipy==1.5.4
seaborn==0.11.2
Send2Trash==1.5.0
sentry-sdk==1.29.2
setproctitle==1.3.2
setuptools==61.2.0
simplejson==3.17.6
six==1.16.0
sklearn==0.0
smmap==5.0.0
SQLAlchemy==1.4.48
tabulate==0.8.7
tbb==2021.6.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
tensorpack==0.10.1
termcolor==1.1.0
terminado==0.8.1
terminaltables==3.1.10
testpath==0.4.2
thop==0.0.31.post2005241907
threadpoolctl==2.1.0
tifffile==2021.11.2
timesformer==1.0
timm==0.5.4
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.9.0
torch==1.8.1+cu111
torchaudio==0.8.1
torchinfo==1.8.0
torchmetrics==0.11.4
torchnet==0.0.4
torchpack==0.2.2
torchstat==0.0.7
torchsummary==1.5.1
torchvideotransforms==0.1.2
torchvision==0.9.1+cu111
tornado==5.1
tqdm==4.64.0
traitlets==4.3.2
transaction==3.1.0
transformers==4.30.2
translationstring==1.4
typing_extensions==4.4.0
ultralytics==8.0.145
urllib3==1.26.16
velruse==1.1.1
venusian==3.0.0
visdom==0.2.3
wandb==0.15.8
wcwidth==0.2.6
webencodings==0.5.1
WebOb==1.8.7
websocket-client==1.4.2
Werkzeug==1.0.1
wheel==0.37.1
widgetsnbextension==3.4.2
wrapt==1.15.0
WTForms==3.0.1
wtforms-recaptcha==0.3.2
xlrd==2.0.1
yacs==0.1.8
yapf==0.32.0
zipp==3.4.0
zope.deprecation==5.0
zope.interface==6.0
zope.sqlalchemy==2.0
